@@ -1,9 +1,9 @@
 import './styles/app.scss';
 import  'mapbox-gl/src/css/mapbox-gl.css'
-
-// start the Stimulus application
 import './bootstrap'
 import { Tooltip, Carousel, Popover } from 'bootstrap'
+import Masonry from 'masonry-layout'
+import Lightpick from 'lightpick'
 
 let tooltipTriggerList = [].slice.call(document.querySelectorAll('[data-bs-toggle="tooltip"]'))
 let tooltipList = tooltipTriggerList.map(function (tooltipTriggerEl) {
@@ -20,4 +20,25 @@ let carousel = document.querySelector('.carousel.preview-carousel')
   interval: false,
   wrap: true,
   ride: false
+})
+
+var elem = document.querySelector('.grid');
+let masonry = new Masonry(elem, {
+  itemSelector: '.grid-item',
+});
+
+document.addEventListener('DOMContentLoaded', () => {
+  let date = new Date()
+  var picker = new Lightpick({
+    field: document.querySelector('.pl-datepicker'),
+    secondField: document.querySelector('.pl-datepicker-second-input'),
+    singleDate: false,
+    minDate: new Date(),
+    onSelect: function (start, end) {
+      var str = '';
+      str += start ? start.format('Do MMMM YYYY') + ' to ' : '';
+      str += end ? end.format('Do MMMM YYYY') : '...';
+      document.getElementById('result-3').innerHTML = str;
+    }
+  });
 })
