@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
 use Rector\Core\ValueObject\PhpVersion;
+use Rector\Set\ValueObject\LevelSetList;
+use Rector\Symfony\Set\SymfonySetList;
 
 return static function (RectorConfig $rectorConfig) {
     $rectorConfig->importNames();
@@ -13,9 +15,10 @@ return static function (RectorConfig $rectorConfig) {
         __DIR__ . '/tests',
     ]);
     $rectorConfig->phpstanConfig(__DIR__ . '/phpstan.neon');
-    $rectorConfig->phpVersion(PhpVersion::PHP_74);
+    $rectorConfig->phpVersion(PhpVersion::PHP_81);
 
-    // $rectorConfig->import(DoctrineSetList::DOCTRINE_REPOSITORY_AS_SERVICE);
+     $rectorConfig->import(SymfonySetList::SYMFONY_60);
+//     $rectorConfig->import(DoctrineSetList::DOCTRINE_REPOSITORY_AS_SERVICE);
     // $rectorConfig->import(SetList::CODE_QUALITY);
     // $rectorConfig->import(SetList::PHP_81);
     // $rectorConfig->import(SetList::TYPE_DECLARATION);
