@@ -4,71 +4,72 @@ namespace App\Controller;
 
 use App\Entity\Asset;
 use App\Entity\Image;
+use App\Factory\PublicationFactory;
+use App\Form\AssetSearchFormType;
 use App\Form\AssetType;
 use App\Repository\AssetRepository;
 use App\Repository\ImageRepository;
-use App\Serializer\AssetNormalizer;
+use App\Repository\PublicationRepository;
 use App\Service\Base64FileUtil;
+use App\Service\ImageUploadService;
+use App\Service\RegionalSettingsService\RegionalSettingsService;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Mercure\HubInterface;
+use Symfony\Component\Mercure\Update;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Serializer\Encoder\JsonEncoder;
-use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
-use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
-use Symfony\Component\Serializer\Serializer;
-use Twig\Environment;
+use Symfony\Component\Serializer\SerializerInterface;
 
-#[Route('/{_locale}/')]
+#[Route('/properties')]
 class AssetController extends AbstractController
 {
     public function __construct(
-        private AssetRepository $assetRepository,
-        private Base64FileUtil $base64FileUtil,
-        private ImageRepository $imageRepository,
-        private Environment $twig,
+        private readonly AssetRepository $assetRepository,
+        private readonly ImageRepository $imageRepository,
+        private readonly SerializerInterface $serializer,
+        private readonly PublicationFactory $publicationFactory,
+        private readonly PublicationRepository $publicationRepository,
+        private readonly ImageUploadService $imageUploadService,
     ) {
     }
 
-    #[Route(path: '', name: 'app_asset_index', methods: ['GET'])]
+    #[Route(path: '/', name: 'app_asset_index', methods: ['GET'])]
     public function index(Request $request): Response
     {
-        /** @var string $type */
-        $type = $request->get('type');
-        /** @var string $term */
-        $term = $request->get('term');
-        /** @var int $minSqm */
-        $minSqm = $request->get('min-sqm');
-        /** @var int $maxSqm */
-        $maxSqm = $request->get('max-sqm');
-        /** @var int $minPrice */
-        $minPrice = $request->get('min-price');
-        /** @var int $maxPrice */
-        $maxPrice = $request->get('max-price');
+        $searchForm = $this->createForm(AssetSearchFormType::class);
+        $searchForm->handleRequest($request);
+        $assets = $this->assetRepository->findAll();
 
-        $assets = $this->assetRepository->findBySearch($minSqm, $maxSqm, $minPrice, $maxPrice, $type, $term);
+        $json = new JsonResponse($this->serializer->serialize($assets, 'json', ['groups' => 'asset_map']));
 
-        $objectNormalizer = new ObjectNormalizer();
-        $serializer = new Serializer([
-            new AssetNormalizer($objectNormalizer, $this->twig),
-        ], [
-            new JsonEncoder()
-        ]);
+        if ($searchForm->isSubmitted() && $searchForm->isValid()) {
 
-        $features = $serializer->serialize([
-            'type' => 'FeatureCollection',
-            'features' => $assets
-        ], 'json', [AbstractNormalizer::IGNORED_ATTRIBUTES => ['id', 'createdAt', 'images', 'owner', 'reviews']]);
+            $assets = $this->handleSearchForm($request, $searchForm);
+            $json = new JsonResponse($this->serializer->serialize($assets, 'json', ['groups' => 'asset_map']));
+
+            return $this->render('asset/index.html.twig', [
+                'features' => $json->getContent(),
+                'assets' => $assets,
+                'searchForm' => $searchForm->createView()
+            ]);
+
+        }
 
         return $this->render('asset/index.html.twig', [
+            'features' => $json->getContent(),
             'assets' => $assets,
-            'features' => $features
+            'searchForm' => $searchForm->createView()
         ]);
     }
 
-    #[Route('new', name: 'app_asset_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, AssetRepository $assetRepository): Response
+    #[Route(path: '/new', name: 'app_asset_new', methods: ['GET', 'POST'])]
+    #[IsGranted('ROLE_USER')]
+    public function new(Request $request): Response
     {
         $asset = new Asset();
         $asset->setOwner($this->getUser());
@@ -81,14 +82,20 @@ class AssetController extends AbstractController
             $files = $form->get('images')->getData();
 
             foreach ($files as $file) {
+                $info = getimagesize($file);
+                [$x, $y] = $info;
 
-                $fileContentBase64 = $this->base64FileUtil->encodeBase64($file->getContent());
+                $fileContentBase64 = $this->imageUploadService->process($file, 800, true);
                 $image = new Image($fileContentBase64, $asset);
+                $image->setHeight($y);
+                $image->setWidth($x);
                 $this->imageRepository->add($image);
             }
 
-            $assetRepository->add($asset);
-            return $this->redirectToRoute('user_account', [], Response::HTTP_SEE_OTHER);
+            $publication = $this->publicationFactory->create($asset, null);
+            $this->publicationRepository->add($publication);
+            $this->assetRepository->add($asset);
+            return $this->redirectToRoute('user_assets');
         }
 
         return $this->renderForm('asset/new.html.twig', [
@@ -97,7 +104,7 @@ class AssetController extends AbstractController
         ]);
     }
 
-    #[Route('images/{id}', name: 'asset_images', methods: ['GET'])]
+    #[Route(path: '/images/{id}', name: 'asset_images', methods: ['GET'])]
     public function images(Asset $asset): Response
     {
         return $this->render('asset/images/show.html.twig', [
@@ -105,15 +112,16 @@ class AssetController extends AbstractController
         ]);
     }
 
-    #[Route('show/{id}', name: 'app_asset_show', methods: ['GET'])]
+    #[Route('/show/{id}', name: 'app_asset_show', methods: ['GET'])]
     public function show(Asset $asset): Response
     {
+
         return $this->render('asset/show.html.twig', [
             'asset' => $asset
         ]);
     }
 
-    #[Route('edit/{id}', name: 'app_asset_edit', methods: ['GET', 'POST'])]
+    #[Route(path: '/edit/{id}', name: 'app_asset_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Asset $asset, AssetRepository $assetRepository): Response
     {
         $form = $this->createForm(AssetType::class, $asset);
@@ -131,16 +139,33 @@ class AssetController extends AbstractController
         ]);
     }
 
-    #[Route('delete/{id}', name: 'app_asset_delete', methods: ['POST'])]
+    #[Route(path: '/delete/{id}', name: 'app_asset_delete', methods: ['POST'])]
     public function delete(Request $request, Asset $asset, AssetRepository $assetRepository): Response
     {
-        if (
-            $this->isCsrfTokenValid('delete' . $asset->getId(),
-                (string) $request->request->get('_token')))
-        {
+        if ($this->isCsrfTokenValid(
+            'delete' . $asset->getId(),
+            (string)$request->request->get('_token')
+        )) {
             $assetRepository->remove($asset);
         }
 
         return $this->redirectToRoute('app_asset_index', [], Response::HTTP_SEE_OTHER);
+    }
+
+    private function handleSearchForm(Request $request, FormInterface $searchForm): mixed
+    {
+        return $this->assetRepository->findBySearch(
+            minSqm: $searchForm->get('minSqm')->getData(),
+            maxSqm: $searchForm->get('maxSqm')->getData(),
+            minPrice: $searchForm->get('minPrice')->getData(),
+            maxPrice: $searchForm->get('maxPrice')->getData(),
+            type: $searchForm->get('type')->getData(),
+            term: $searchForm->get('term')->getData(),
+            userType: $searchForm->get('userType')->getData(),
+            title: $searchForm->get('title')->getData(),
+            floor: $searchForm->get('floor')->getData(),
+            agencyFee: $searchForm->get('agencyFee')->getData(),
+            address: $searchForm->get('address')->getData()
+        );
     }
 }
