@@ -7,25 +7,41 @@ use App\Form\RegistrationFormType;
 use App\Security\AppCustomAuthenticator;
 use App\Service\AvatarService;
 use Doctrine\ORM\EntityManagerInterface;
+use Psr\Log\LoggerInterface;
+use Symfony\Bridge\Twig\Mime\TemplatedEmail;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
+use Symfony\Component\Mailer\MailerInterface;
+use Symfony\Component\Mime\Address;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Security\Http\Authentication\UserAuthenticatorInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
+#[Route('/security')]
 class RegistrationController extends AbstractController
 {
     public function __construct(
-        private AvatarService $avatarService
-    )
-    {
+        private readonly AvatarService $avatarService,
+        private readonly TranslatorInterface $translator,
+        private readonly UrlGeneratorInterface $urlGenerator,
+        private readonly MailerInterface $mailer,
+        private readonly LoggerInterface $logger
+    ) {
     }
 
+
     #[Route('/register', name: 'app_register')]
-    public function register(Request $request, UserPasswordHasherInterface $userPasswordHasher, UserAuthenticatorInterface $userAuthenticator, AppCustomAuthenticator $authenticator, EntityManagerInterface $entityManager): ?Response
-    {
+    public function register(
+        Request $request,
+        UserPasswordHasherInterface $userPasswordHasher,
+        UserAuthenticatorInterface $userAuthenticator,
+        AppCustomAuthenticator $authenticator,
+        EntityManagerInterface $entityManager
+    ): ?Response {
         $user = new User();
         $form = $this->createForm(RegistrationFormType::class, $user);
         $form->handleRequest($request);
@@ -35,15 +51,16 @@ class RegistrationController extends AbstractController
             $user->setAvatar($this->avatarService->createAvatar($user->getEmail()));
 
             $user->setPassword(
-            $userPasswordHasher->hashPassword(
+                $userPasswordHasher->hashPassword(
                     $user,
-                 $form->get('plainPassword')->getData()
+                    $form->get('plainPassword')->getData()
                 )
             );
 
             $entityManager->persist($user);
             $entityManager->flush();
-            // do anything else you need here, like send an email
+
+            $this->sendConfirmationEmail($user);
 
             return $userAuthenticator->authenticateUser(
                 $user,
@@ -55,5 +72,28 @@ class RegistrationController extends AbstractController
         return $this->render('registration/register.html.twig', [
             'registrationForm' => $form->createView(),
         ]);
+    }
+
+    private function sendConfirmationEmail(
+        User $user
+    ): void {
+        $email = new TemplatedEmail();
+        $email->from('fabien@example.com');
+        $email->to(new Address($user->getEmail()));
+        $email->subject(
+            $this->translator->trans('email.confirmation.subject')
+        );
+        $email->htmlTemplate('emails/confirm-email-address.html.twig');
+        $email->context([
+            'user' => $user,
+            'path' => $this->urlGenerator->generate('app_account_confirm', ['token' => $user->getToken()]),
+        ]);
+
+        try {
+            $this->mailer->send($email);
+        } catch (TransportExceptionInterface $exception) {
+            $this->logger->log(null, $exception->getMessage());
+        }
+
     }
 }
