@@ -2,67 +2,78 @@
 
 namespace App\Entity;
 
+use Stringable;
 use App\Repository\AssetRepository;
+use Carbon\Carbon;
 use DateTimeImmutable;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Bridge\Doctrine\IdGenerator\UuidGenerator;
 use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Component\Serializer\Annotation\Groups;
 use Symfony\Component\Uid\Uuid;
 
 #[ORM\Entity(repositoryClass: AssetRepository::class)]
-class Asset
+class Asset implements Stringable
 {
     #[ORM\Id]
     #[ORM\GeneratedValue(strategy: 'CUSTOM')]
     #[ORM\Column(type: 'uuid', unique: true)]
     #[ORM\CustomIdGenerator(UuidGenerator::class)]
+    #[Groups(['asset_map'])]
     private Uuid $id;
 
     #[ORM\Column(type: 'string', length: 255)]
+    #[Groups(['asset_map'])]
     private string $title;
 
     #[ORM\Column(type: 'text', nullable: true)]
     private ?string $description = null;
 
     #[ORM\Column(type: 'integer')]
+    #[Groups(['asset_map'])]
     private int $sqm;
 
     #[ORM\Column(type: 'float')]
+    #[Groups(['asset_map'])]
     private float $longitude = 0.00;
 
     #[ORM\Column(type: 'float')]
+    #[Groups(['asset_map'])]
     private float $latitude = 0.00;
 
     #[ORM\Column(type: 'string', length: 10)]
+    #[Groups(['asset_map'])]
     private string $type;
 
     #[ORM\Column(type: 'string', length: 255)]
+    #[Groups(['asset_map'])]
     private string $term;
 
     #[ORM\ManyToOne(targetEntity: User::class, inversedBy: 'assets')]
     private null|User|UserInterface $owner = null;
 
     #[ORM\Column(type: 'datetime_immutable')]
+    #[Groups(['asset_map'])]
     private DateTimeImmutable $createdAt;
 
     /**
      * @var ArrayCollection<int,Image>
      */
-    #[ORM\OneToMany(mappedBy: 'asset', targetEntity: Image::class)]
+    #[ORM\OneToMany(mappedBy: 'asset', targetEntity: Image::class, cascade: ['persist'])]
     private Collection $images;
 
-    #[ORM\Column(type: 'boolean')]
-    private bool $isPublished = false;
-
     #[ORM\Column(type: 'integer')]
-    private int $fee = 1;
+    #[Groups(['asset_map'])]
+    private ?int $price;
 
     #[ORM\Column(type: 'string', length: 255)]
+    #[Groups(['asset_map'])]
     private string $address;
 
     #[ORM\Column(type: 'integer')]
+    #[Groups(['asset_map'])]
     private int $floor;
 
     /**
@@ -71,11 +82,37 @@ class Asset
     #[ORM\OneToMany(mappedBy: 'asset', targetEntity: Review::class)]
     private Collection $reviews;
 
-    public function __construct()
-    {
+    #[ORM\OneToMany(mappedBy: 'asset', targetEntity: Reaction::class)]
+    private Collection $reactions;
+
+    #[ORM\OneToMany(mappedBy: 'asset', targetEntity: AssetView::class)]
+    private Collection $views;
+
+    #[ORM\OneToMany(mappedBy: 'asset', targetEntity: Publication::class, cascade: ['persist'])]
+    private Collection $publications;
+
+    #[ORM\Column(type: 'integer', nullable: true)]
+    private ?int $agencyFee;
+
+    #[ORM\OneToMany(mappedBy: 'asset', targetEntity: Bookmark::class)]
+    private Collection $bookmarks;
+
+    #[ORM\OneToMany(mappedBy: 'asset', targetEntity: Tender::class)]
+    private Collection $tenders;
+
+    #[ORM\OneToOne(targetEntity: Tender::class, cascade: ['persist', 'remove'])]
+    #[ORM\JoinColumn(nullable: true)]
+    private ?Tender $tender = null;
+
+    public function __construct() {
         $this->createdAt = new DateTimeImmutable();
         $this->images = new ArrayCollection();
         $this->reviews = new ArrayCollection();
+        $this->reactions = new ArrayCollection();
+        $this->views = new ArrayCollection();
+        $this->publications = new ArrayCollection();
+        $this->bookmarks = new ArrayCollection();
+        $this->tenders = new ArrayCollection();
     }
 
     public function getId(): Uuid
@@ -222,28 +259,15 @@ class Asset
         return $this;
     }
 
-    public function getIsPublished(): bool
+    public function getPrice(): int
     {
-        return $this->isPublished;
+        return $this->price / 100;
     }
 
-    public function setIsPublished(bool $isPublished): self
+    public function setPrice(int $price): int
     {
-        $this->isPublished = $isPublished;
-
-        return $this;
-    }
-
-    public function getFee(): int
-    {
-        return $this->fee;
-    }
-
-    public function setFee(int $fee): self
-    {
-        $this->fee = $fee;
-
-        return $this;
+        $this->price = $price * 100;
+        return $this->price;
     }
 
     public function getAddress(): string
@@ -297,4 +321,283 @@ class Asset
 
         return $this;
     }
+
+    /**
+     * @return Collection<int, Reaction>
+     */
+    public function getReactions(): Collection
+    {
+        return $this->reactions;
+    }
+
+    public function addReaction(Reaction $reaction): self
+    {
+        if (!$this->reactions->contains($reaction)) {
+            $this->reactions[] = $reaction;
+            $reaction->setAsset($this);
+        }
+
+        return $this;
+    }
+
+    public function removeReaction(Reaction $reaction): self
+    {
+        if ($this->reactions->removeElement($reaction)) {
+            // set the owning side to null (unless already changed)
+            if ($reaction->getAsset() === $this) {
+                $reaction->setAsset(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function getLikes(): int
+    {
+        $likes = 0;
+        /**
+         * @var Reaction $reaction
+         */
+        foreach ($this->reactions as $reaction) {
+            if ($reaction->isLike()) {
+                $likes++;
+            }
+        }
+
+        return $likes;
+    }
+
+    public function getDislikes(): int
+    {
+        $dislikes = 0;
+        /**
+         * @var Reaction $reaction
+         */
+        foreach ($this->reactions as $reaction) {
+            if ($reaction->isDislike()) {
+                $dislikes++;
+            }
+        }
+
+        return $dislikes;
+    }
+
+    public function __toString(): string
+    {
+        return $this->title;
+    }
+
+    public function isNew(): bool
+    {
+        $week = Carbon::parse($this->createdAt)->addWeek();
+        return Carbon::now()->isBefore($week);
+    }
+
+    public function isLikedByUser(User $user): bool
+    {
+        /** @var Reaction $reaction */
+        foreach ($this->reactions as $reaction) {
+            if ($reaction->getType() == 'like' && $reaction->getOwner() === $user) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    public function isDislikedByUser(User $user): bool
+    {
+        /** @var Reaction $reaction */
+        foreach ($this->reactions as $reaction) {
+            if ($reaction->getType() == 'dislike' && $reaction->getOwner() === $user) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * @return Collection<int, AssetView>
+     */
+    public function getViews(): Collection
+    {
+        return $this->views;
+    }
+
+    public function addView(AssetView $view): self
+    {
+        if (!$this->views->contains($view)) {
+            $this->views[] = $view;
+            $view->setAsset($this);
+        }
+
+        return $this;
+    }
+
+    public function removeView(AssetView $view): self
+    {
+        if ($this->views->removeElement($view)) {
+            // set the owning side to null (unless already changed)
+            if ($view->getAsset() === $this) {
+                $view->setAsset(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function getPageViews(): int
+    {
+        $views = 0;
+        /** @var AssetView $view */
+        foreach ($this->views as $view) {
+            if (!$view->hasUser()) {
+                $views++;
+            }
+        }
+
+        return $views;
+    }
+
+    /**
+     * @return Collection<int, Publication>
+     */
+    public function getPublications(): Collection
+    {
+        return $this->publications;
+    }
+
+    public function addPublication(Publication $publication): self
+    {
+        if (!$this->publications->contains($publication)) {
+            $this->publications[] = $publication;
+            $publication->setAsset($this);
+        }
+
+        return $this;
+    }
+
+    public function removePublication(Publication $publication): self
+    {
+        if ($this->publications->removeElement($publication)) {
+            // set the owning side to null (unless already changed)
+            if ($publication->getAsset() === $this) {
+                $publication->setAsset(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function getIsPublished(): bool
+    {
+        /** @var Publication $publication */
+        foreach ($this->publications as $publication){
+
+            if(!$publication->getIsActive()){
+                continue;
+            }
+
+            return true;
+        }
+
+        return false;
+    }
+
+    public function getAgencyFee(): ?int
+    {
+        return $this->agencyFee;
+    }
+
+    public function setAgencyFee(?int $agencyFee): self
+    {
+        $this->agencyFee = $agencyFee;
+        return $this;
+    }
+
+    /**
+     * @return Collection<int, Bookmark>
+     */
+    public function getBookmarks(): Collection
+    {
+        return $this->bookmarks;
+    }
+
+    public function addBookmark(Bookmark $bookmark): self
+    {
+        if (!$this->bookmarks->contains($bookmark)) {
+            $this->bookmarks[] = $bookmark;
+            $bookmark->setAsset($this);
+        }
+
+        return $this;
+    }
+
+    public function removeBookmark(Bookmark $bookmark): self
+    {
+        if ($this->bookmarks->removeElement($bookmark)) {
+            // set the owning side to null (unless already changed)
+            if ($bookmark->getAsset() === $this) {
+                $bookmark->setAsset(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function isBookmarkedByUser(User $user): bool
+    {
+        /** @var Bookmark $bookmark */
+        foreach ($this->getBookmarks() as $bookmark) {
+            if ($bookmark->getOwner() === $user ) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * @return Collection<int, Tender>
+     */
+    public function getTenders(): Collection
+    {
+        return $this->tenders;
+    }
+
+    public function addTender(Tender $tender): self
+    {
+        if (!$this->tenders->contains($tender)) {
+            $this->tenders[] = $tender;
+            $tender->setAsset($this);
+        }
+
+        return $this;
+    }
+
+    public function removeTender(Tender $tender): self
+    {
+        if ($this->tenders->removeElement($tender)) {
+            // set the owning side to null (unless already changed)
+            if ($tender->getAsset() === $this) {
+                $tender->setAsset(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function getTender(): ?Tender
+    {
+        return $this->tender;
+    }
+
+    public function setTender(?Tender $tender): self
+    {
+        $this->tender = $tender;
+
+        return $this;
+    }
+
 }
