@@ -2,6 +2,7 @@
 
 namespace App\Entity;
 
+use Stringable;
 use App\Repository\ImageRepository;
 use DateTimeImmutable;
 use Doctrine\ORM\Mapping as ORM;
@@ -9,7 +10,7 @@ use Symfony\Bridge\Doctrine\IdGenerator\UuidGenerator;
 use Symfony\Component\Uid\Uuid;
 
 #[ORM\Entity(repositoryClass: ImageRepository::class)]
-class Image
+class Image implements Stringable
 {
     #[ORM\Id]
     #[ORM\GeneratedValue(strategy: 'CUSTOM')]
@@ -17,23 +18,21 @@ class Image
     #[ORM\CustomIdGenerator(UuidGenerator::class)]
     private Uuid $id;
 
-    #[ORM\Column(type: 'text')]
-    private string $base64;
-
-    #[ORM\ManyToOne(targetEntity: Asset::class, cascade: ['persist'], inversedBy: 'images')]
-    private ?Asset $asset;
-
     #[ORM\Column(type: 'datetime_immutable')]
     private DateTimeImmutable $createdAt;
 
+    #[ORM\Column(type: 'integer', nullable: true)]
+    private $height;
+
+    #[ORM\Column(type: 'integer', nullable: true)]
+    private $width;
+
     public function __construct(
-        string $base64,
-        ?Asset $asset
+        #[ORM\Column(type: 'text')] private string $base64,
+        #[ORM\ManyToOne(targetEntity: Asset::class, cascade: ['persist'], inversedBy: 'images')] private ?Asset $asset
 
     )
     {
-        $this->base64 = $base64;
-        $this->asset = $asset;
         $this->createdAt = new DateTimeImmutable();
     }
 
@@ -74,6 +73,35 @@ class Image
     public function setCreatedAt(\DateTimeImmutable $createdAt): self
     {
         $this->createdAt = $createdAt;
+
+        return $this;
+    }
+
+    public function __toString() : string
+    {
+        return $this->base64;
+    }
+
+    public function getHeight(): ?int
+    {
+        return $this->height;
+    }
+
+    public function setHeight(?int $height): self
+    {
+        $this->height = $height;
+
+        return $this;
+    }
+
+    public function getWidth(): ?int
+    {
+        return $this->width;
+    }
+
+    public function setWidth(?int $width): self
+    {
+        $this->width = $width;
 
         return $this;
     }
