@@ -2,6 +2,7 @@
 
 namespace App\Entity;
 
+use Stringable;
 use App\Repository\UserRepository;
 use DateTimeImmutable;
 use Doctrine\Common\Collections\ArrayCollection;
@@ -16,71 +17,73 @@ use Symfony\Component\Uid\Uuid;
 #[ORM\Entity(repositoryClass: UserRepository::class)]
 #[ORM\Table(name: '`user`')]
 #[UniqueEntity(fields: ['email'], message: 'There is already an account with this email')]
-class User implements UserInterface, PasswordAuthenticatedUserInterface
+class User implements UserInterface, PasswordAuthenticatedUserInterface, Stringable
 {
     #[ORM\Id]
     #[ORM\GeneratedValue(strategy: 'CUSTOM')]
     #[ORM\Column(type: 'uuid', unique: true)]
     #[ORM\CustomIdGenerator(UuidGenerator::class)]
     private Uuid $id;
-
     #[ORM\Column(type: 'string', length: 180, unique: true)]
     private string $email;
-
     #[ORM\Column(type: 'json')]
     private array $roles = [];
-
     #[ORM\Column(type: 'string')]
     private string $password;
-
     #[ORM\Column(type: 'text')]
     private string $avatar;
-
     /**
      * @var ArrayCollection<int, Message>
      */
     #[ORM\OneToMany(mappedBy: 'owner', targetEntity: Message::class)]
     private Collection $messages;
-
     #[ORM\Column(type: 'datetime_immutable')]
     private DateTimeImmutable $createdAt;
-
     /**
      * @var ArrayCollection<int, Asset>
      */
     #[ORM\OneToMany(mappedBy: 'owner', targetEntity: Asset::class)]
     private Collection $assets;
-
     #[ORM\Column(type: 'string', length: 255)]
     private ?string $firstName = null;
-
     #[ORM\Column(type: 'string', length: 255)]
     private ?string $lastName = null;
-
     /**
      * @var ArrayCollection<int, Conversation>
      */
-    #[ORM\ManyToMany(targetEntity: Conversation::class, mappedBy: 'users')]
+    #[ORM\ManyToMany(targetEntity: Conversation::class, mappedBy: 'users', cascade: ['persist'])]
     private Collection $conversations;
-
     /**
      * @var ArrayCollection<int, Insight>
      */
     #[ORM\OneToMany(mappedBy: 'owner', targetEntity: Insight::class)]
     private Collection $insights;
-
     /**
      * @var ArrayCollection<int, Review>
      */
-    #[ORM\OneToMany(mappedBy: 'owner', targetEntity: Review::class)]
+    #[ORM\OneToMany(mappedBy: 'author', targetEntity: Review::class,)]
     private Collection $ownedReviews;
-
     /**
      * @var ArrayCollection<int, Review>
      */
-    #[ORM\OneToMany(mappedBy: 'reviewee', targetEntity: Review::class)]
+    #[ORM\OneToMany(mappedBy: 'user', targetEntity: Review::class)]
     private Collection $reviews;
-
+    #[ORM\OneToMany(mappedBy: 'owner', targetEntity: Reaction::class)]
+    private Collection $reactions;
+    #[ORM\OneToMany(mappedBy: 'owner', targetEntity: AssetView::class)]
+    private Collection $assetViews;
+    #[ORM\Column(type: 'uuid', unique: true)]
+    private Uuid $token;
+    #[ORM\Column(type: 'boolean')]
+    private bool $isEnabled = false;
+    #[ORM\OneToMany(mappedBy: 'owner', targetEntity: SocialAuth::class)]
+    private Collection $socialAuths;
+    #[ORM\OneToOne(mappedBy: 'owner', targetEntity: Agency::class, cascade: ['persist', 'remove'])]
+    private ?Agency $agency = null;
+    #[ORM\OneToMany(mappedBy: 'owner', targetEntity: Bookmark::class)]
+    private Collection $bookmarks;
+    #[ORM\OneToMany(mappedBy: 'owner', targetEntity: Bid::class)]
+    private Collection $bids;
     public function __construct()
     {
         $this->messages = new ArrayCollection();
@@ -90,25 +93,27 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
         $this->insights = new ArrayCollection();
         $this->ownedReviews = new ArrayCollection();
         $this->reviews = new ArrayCollection();
+        $this->reactions = new ArrayCollection();
+        $this->assetViews = new ArrayCollection();
+        $this->token = Uuid::v4();
+        $this->socialAuths = new ArrayCollection();
+        $this->bookmarks = new ArrayCollection();
+        $this->bids = new ArrayCollection();
     }
-
     public function getId(): Uuid
     {
         return $this->id;
     }
-
     public function getEmail(): string
     {
         return $this->email;
     }
-
     public function setEmail(string $email): self
     {
         $this->email = $email;
 
         return $this;
     }
-
     /**
      * A visual identifier that represents this user.
      *
@@ -116,12 +121,11 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
      */
     public function getUserIdentifier(): string
     {
-        return (string)$this->email;
+        return (string) $this->email;
     }
-
     /**
-     * @see UserInterface
      * @return mixed[]
+     * @see UserInterface
      */
     public function getRoles(): array
     {
@@ -131,7 +135,6 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
 
         return array_unique($roles);
     }
-
     /**
      * @param mixed[] $roles
      */
@@ -141,7 +144,6 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
 
         return $this;
     }
-
     /**
      * @see PasswordAuthenticatedUserInterface
      */
@@ -149,14 +151,12 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     {
         return $this->password;
     }
-
     public function setPassword(string $password): self
     {
         $this->password = $password;
 
         return $this;
     }
-
     /**
      * @see UserInterface
      */
@@ -165,19 +165,16 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
         // If you store any temporary, sensitive data on the user, clear it here
         // $this->plainPassword = null;
     }
-
     public function getAvatar(): string
     {
         return $this->avatar;
     }
-
     public function setAvatar(string $avatar): self
     {
         $this->avatar = $avatar;
 
         return $this;
     }
-
     /**
      * @return Collection<int, Message>
      */
@@ -185,17 +182,15 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     {
         return $this->messages;
     }
-
     public function addMessage(Message $message): self
     {
-        if (!$this->messages->contains($message)) {
+        if (! $this->messages->contains($message)) {
             $this->messages[] = $message;
             $message->setOwner($this);
         }
 
         return $this;
     }
-
     public function removeMessage(Message $message): self
     {
         // set the owning side to null (unless already changed)
@@ -205,19 +200,16 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
 
         return $this;
     }
-
     public function getCreatedAt(): DateTimeImmutable
     {
         return $this->createdAt;
     }
-
     public function setCreatedAt(DateTimeImmutable $createdAt): self
     {
         $this->createdAt = $createdAt;
 
         return $this;
     }
-
     /**
      * @return ArrayCollection<int, Asset>
      */
@@ -225,17 +217,15 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     {
         return $this->assets;
     }
-
     public function addAsset(Asset $asset): self
     {
-        if (!$this->assets->contains($asset)) {
+        if (! $this->assets->contains($asset)) {
             $this->assets[] = $asset;
             $asset->setOwner($this);
         }
 
         return $this;
     }
-
     public function removeAsset(Asset $asset): self
     {
         // set the owning side to null (unless already changed)
@@ -245,41 +235,34 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
 
         return $this;
     }
-
     public function getFullName(): string
     {
         return $this->firstName . " " . $this->lastName;
     }
-
     public function getFirstName(): ?string
     {
         return $this->firstName;
     }
-
     public function setFirstName(string $firstName): self
     {
         $this->firstName = $firstName;
 
         return $this;
     }
-
     public function getLastName(): ?string
     {
         return $this->lastName;
     }
-
     public function setLastName(string $lastName): self
     {
         $this->lastName = $lastName;
 
         return $this;
     }
-
     public function __toString(): string
     {
         return $this->getFullName();
     }
-
     /**
      * @return ArrayCollection<int, Conversation>
      */
@@ -287,17 +270,15 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     {
         return $this->conversations;
     }
-
     public function addConversation(Conversation $conversation): self
     {
-        if (!$this->conversations->contains($conversation)) {
+        if (! $this->conversations->contains($conversation)) {
             $this->conversations[] = $conversation;
             $conversation->addUser($this);
         }
 
         return $this;
     }
-
     public function removeConversation(Conversation $conversation): self
     {
         if ($this->conversations->removeElement($conversation)) {
@@ -306,7 +287,6 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
 
         return $this;
     }
-
     /**
      * @return ArrayCollection<int, Insight>
      */
@@ -314,17 +294,15 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     {
         return $this->insights;
     }
-
     public function addInsight(Insight $insight): self
     {
-        if (!$this->insights->contains($insight)) {
+        if (! $this->insights->contains($insight)) {
             $this->insights[] = $insight;
             $insight->setOwner($this);
         }
 
         return $this;
     }
-
     public function removeInsight(Insight $insight): self
     {
         // set the owning side to null (unless already changed)
@@ -334,7 +312,6 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
 
         return $this;
     }
-
     /**
      * @return ArrayCollection<int, Review>
      */
@@ -342,27 +319,24 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     {
         return $this->ownedReviews;
     }
-
     public function addOwnedReview(Review $ownedReview): self
     {
-        if (!$this->ownedReviews->contains($ownedReview)) {
+        if (! $this->ownedReviews->contains($ownedReview)) {
             $this->ownedReviews[] = $ownedReview;
-            $ownedReview->setOwner($this);
+            $ownedReview->setAuthor($this);
         }
 
         return $this;
     }
-
     public function removeOwnedReview(Review $ownedReview): self
     {
         // set the owning side to null (unless already changed)
-        if ($this->ownedReviews->removeElement($ownedReview) && $ownedReview->getOwner() === $this) {
-            $ownedReview->setOwner(null);
+        if ($this->ownedReviews->removeElement($ownedReview) && $ownedReview->getAuthor() === $this) {
+            $ownedReview->setAuthor(null);
         }
 
         return $this;
     }
-
     /**
      * @return ArrayCollection<int, Review>
      */
@@ -370,25 +344,197 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     {
         return $this->reviews;
     }
-
     public function addReview(Review $review): self
     {
-        if (!$this->reviews->contains($review)) {
+        if (! $this->reviews->contains($review)) {
             $this->reviews[] = $review;
-            $review->setReviewee($this);
+            $review->setUser($this);
         }
 
         return $this;
     }
-
     public function removeReview(Review $review): self
     {
         // set the owning side to null (unless already changed)
-        if ($this->reviews->removeElement($review) && $review->getReviewee() === $this) {
-            $review->setReviewee(null);
+        if ($this->reviews->removeElement($review) && $review->getUser() === $this) {
+            $review->setUser(null);
         }
 
         return $this;
     }
+    /**
+     * @return Collection<int, Reaction>
+     */
+    public function getReactions(): Collection
+    {
+        return $this->reactions;
+    }
+    public function addReaction(Reaction $reaction): self
+    {
+        if (! $this->reactions->contains($reaction)) {
+            $this->reactions[] = $reaction;
+            $reaction->setOwner($this);
+        }
 
+        return $this;
+    }
+    public function removeReaction(Reaction $reaction): self
+    {
+        if ($this->reactions->removeElement($reaction)) {
+            // set the owning side to null (unless already changed)
+            if ($reaction->getOwner() === $this) {
+                $reaction->setOwner(null);
+            }
+        }
+
+        return $this;
+    }
+    /**
+     * @return Collection<int, AssetView>
+     */
+    public function getAssetViews(): Collection
+    {
+        return $this->assetViews;
+    }
+    public function addAssetView(AssetView $assetView): self
+    {
+        if (! $this->assetViews->contains($assetView)) {
+            $this->assetViews[] = $assetView;
+            $assetView->setOwner($this);
+        }
+
+        return $this;
+    }
+    public function removeAssetView(AssetView $assetView): self
+    {
+        if ($this->assetViews->removeElement($assetView)) {
+            // set the owning side to null (unless already changed)
+            if ($assetView->getOwner() === $this) {
+                $assetView->setOwner(null);
+            }
+        }
+
+        return $this;
+    }
+    public function getToken(): ?Uuid
+    {
+        return $this->token;
+    }
+    public function setToken(?Uuid $token): self
+    {
+        $this->token = $token;
+
+        return $this;
+    }
+    public function isIsEnabled(): ?bool
+    {
+        return $this->isEnabled;
+    }
+    public function setIsEnabled(bool $isEnabled): self
+    {
+        $this->isEnabled = $isEnabled;
+
+        return $this;
+    }
+    /**
+     * @return Collection<int, SocialAuth>
+     */
+    public function getSocialAuths(): Collection
+    {
+        return $this->socialAuths;
+    }
+    public function addSocialAuth(SocialAuth $socialAuth): self
+    {
+        if (! $this->socialAuths->contains($socialAuth)) {
+            $this->socialAuths[] = $socialAuth;
+            $socialAuth->setOwner($this);
+        }
+
+        return $this;
+    }
+    public function removeSocialAuth(SocialAuth $socialAuth): self
+    {
+        if ($this->socialAuths->removeElement($socialAuth)) {
+            // set the owning side to null (unless already changed)
+            if ($socialAuth->getOwner() === $this) {
+                $socialAuth->setOwner(null);
+            }
+        }
+
+        return $this;
+    }
+    public function getAgency(): ?Agency
+    {
+        return $this->agency;
+    }
+    public function setAgency(?Agency $agency): self
+    {
+        // unset the owning side of the relation if necessary
+        if ($agency === null && $this->agency !== null) {
+            $this->agency->setOwner(null);
+        }
+
+        // set the owning side of the relation if necessary
+        if ($agency !== null && $agency->getOwner() !== $this) {
+            $agency->setOwner($this);
+        }
+
+        $this->agency = $agency;
+
+        return $this;
+    }
+    /**
+     * @return Collection<int, Bookmark>
+     */
+    public function getBookmarks(): Collection
+    {
+        return $this->bookmarks;
+    }
+    public function addBookmark(Bookmark $bookmark): self
+    {
+        if (! $this->bookmarks->contains($bookmark)) {
+            $this->bookmarks[] = $bookmark;
+            $bookmark->setOwner($this);
+        }
+
+        return $this;
+    }
+    public function removeBookmark(Bookmark $bookmark): self
+    {
+        if ($this->bookmarks->removeElement($bookmark)) {
+            // set the owning side to null (unless already changed)
+            if ($bookmark->getOwner() === $this) {
+                $bookmark->setOwner(null);
+            }
+        }
+
+        return $this;
+    }
+    /**
+     * @return Collection<int, Bid>
+     */
+    public function getBids(): Collection
+    {
+        return $this->bids;
+    }
+    public function addBid(Bid $bid): self
+    {
+        if (! $this->bids->contains($bid)) {
+            $this->bids[] = $bid;
+            $bid->setOwner($this);
+        }
+
+        return $this;
+    }
+    public function removeBid(Bid $bid): self
+    {
+        if ($this->bids->removeElement($bid)) {
+            // set the owning side to null (unless already changed)
+            if ($bid->getOwner() === $this) {
+                $bid->setOwner(null);
+            }
+        }
+
+        return $this;
+    }
 }
