@@ -11,19 +11,18 @@ use Twig\Environment;
 class AssetNormalizer implements ContextAwareNormalizerInterface
 {
     public function __construct(
-        private ObjectNormalizer $objectNormalizer,
-        private Environment $twig
+        private readonly ObjectNormalizer $objectNormalizer,
+        private readonly Environment $twig
     ) {
     }
 
 
-    /** @param Asset $asset
-     * @return array */
+    /** @param Asset $asset */
     public function normalize(
         mixed $asset,
         string $format = null,
         array $context = []
-    ): array {
+    ) {
 
         $data = $this->objectNormalizer->normalize($asset, 'json', $context);
 
@@ -39,7 +38,7 @@ class AssetNormalizer implements ContextAwareNormalizerInterface
                 'id' => $asset->getId()->toRfc4122(),
                 'title' => $asset->getTitle(),
                 'sqm' => $asset->getSqm(),
-                'fee' => $asset->getFee(),
+                'price' => $asset->getPrice(),
                 'html' => $this->twig->render('asset/_map-asset-preview.html.twig', ['asset' => $asset]),
             ]
         ];
